DeepLab-LargeFOV and reformations
