+++ torchvision_models/segmentation/deeplab.py
@@ -1,3 +1,5 @@
+# Note that this file only implement DeepLabs with ResNet backbone,
+# For the original DeepLab-LargeFOV with VGG backbone, refer to deeplab_vgg.py
 import torch
 from torch import nn
 from torch.nn import functional as F
@@ -9,7 +11,8 @@ __all__ = ["DeepLab"]
 
 
 class DeepLab(_SimpleSegmentationModel):
-    # Should be able to support all deeplab models (>= v2) without CRF
+    # Should be able to support all deeplab models without CRF
+    # Note that v1 is not the official version
     pass
 
 
@@ -29,6 +32,15 @@ class DeepLabV2Head(nn.Sequential):
     def __init__(self, in_channels, num_classes):
         super(DeepLabV2Head, self).__init__(
             ASPP_V2(in_channels, num_classes, [6, 12, 18, 24])
+        )
+
+
+# For better format consistency
+# Not the official VGG backbone version
+class DeepLabV1Head(nn.Sequential):
+    def __init__(self, in_channels, num_classes):
+        super(DeepLabV1Head, self).__init__(
+            LargeFOV(in_channels, num_classes)
         )
 
 
@@ -146,3 +158,12 @@ class ASPP_V2(nn.Module):
         for i in range(len(self.convs) - 1):
             res += self.convs[i + 1](x)
             return res
+
+
+class LargeFOV(nn.Module):
+    def __init__(self, in_channels, num_classes):
+        super(LargeFOV, self).__init__()
+        self.conv = nn.Conv2d(in_channels, num_classes, kernel_size=3, stride=1, padding=12, dilation=12, bias=True)
+
+    def forward(self, x):
+        return self.conv(x)
